Added skeleton to start the SUT using the gym.Env interface
Uses a multiprocessing.Process to start the application. Inside the
child process a QThread listens to click coordinates on a mp.Pipe
connection and sends it to the application which sends an observation,
i.e. a screenshot back.

Alternatively if a boolean is send through the pipe, instead of
coordinates, a random widget of the application will be chosen.