Implemented having a list of currently shown widgets in the different GUI states
On specific button presses the list of currently shown widgets is recreated with the
appropriate widgets. This is done in the main window, as well as in the settings dialog.
When a random click on one of these widgets shall be done, the correct list is chosen, depending
on whether the settings dialog is visible or not.