Added Car Configurator
This is now the third tab in the main window which allows to choose between three cars and select parts for it.
There are tires, interior design and propulsion systems. After selecting them a button can be clicked to print the
configuration. In the settings all parts can be deselected which can lead to deactivating (all) cars.