Fixed clicking function
- Checks now first if a modal window is open, if clicked outside of the modal window discard the click
(as it would with a hardware mouse). Unfortunately QTest.mouseClick() would ignore the modality
- After modal window test, check if a combo box has been opened previously. For that a reference to an opened
combo box is held. A click outside the combo box closes it. This overwrites any click function that has been
set by the modal window test, because clicking on a disabled window while a combo box is opened, also closes
the combo box
- After that the correct click (or no click) is executed while the coverage is measured as before
- After opening or closing a combo box an additional sleep of half a second is done to ensure that the animations
for this opening or closing is done fully. Not doing this would result in several frames where the boxes are half opened