Extended the click function to QActions and fixed a tedious bug when an item of a combo box is selected
- QAction's are not QWidget's, therefore QTest.mouseClick does not work. Instead trigger() is used, a function
of QAction to trigger its action
- A click on an item of an opened QComboBox would not close the combo box, if the events are coming in too fast. This
could also be fixed with adding a half a second sleep in the step() function of the environments, but I opted for adding
the sleep inside the click function and a following QApplication.processEvents() which processes all outstanding events.
This is not done automatically in the click function because it blocks the event loop and the loop continues only after
the click function is returned, except now using QApplication.processEvents() this can be done within the click function.
Nevertheless the half a second sleep is still needed for some reason