Fixed bug when using random widget environment
- Randomly selecting the widgets for a click did work except for QComboBoxes, which would never select anything
different than the first entry. This is again due to the issue that an openened QComboBox is the wrong widget to
select for a click. Instead one has to use the child widget of the openend QComboBox of its view. This has then all
the items that can be clicked.
- The solution is to check if the previously opened QComboBox has been selected again and this selected widget for a click
is then replaced to use the child widget, which can be used for selecting items different to the first item