Fixed bug when clicking; added event filter to indicate last paint event
- Had a bug where clicking inside text/figure printer output had the same object name
as an open combo box. This lead to a failure of the hacks used in the execute_click method
- Now installed an event filter on the whole QApplication that indicates the last paint event, using
a timer. This is useful to have an indication on whether the window is still painted (rendered) or if it
is finished. Unfortunately this is not very reliable, because when the last PaintEvent is registered, it
does not mean that the window is fully painted, because it seems like it then gets sent to the OS which
requires some time to really paint it. Therefore we still sleep for some seconds after the last paint event.