DreamRollout uses SimulatedGUIEnv now, fixed GUIEnv rollout
- Controller now has a predict function that maps its output from [-1, 1] range to integers in [0, 447]
- DreamRollout uses SimulatedGUIEnv internally now. Calls for this in train_controller are updated
- Refactored some constants into own file
- GUIEnvRollout was bugged because the input for the RNN's was not correctly transformed. The workflow now is
to let the controller predict integers in [0, 447] and then transform these integers the same way it is done during
training of the M Model. This can lead to applying tanh and moving to [0, 447] range in controller and then immediately
moving to [-1, 1] range again. This is somewhat unintuitive but should resemble exactly how the M Model is trained and
therefore receive the appropriate input.