MDN Prediction: Fixed a bug where a wrong index was selected due to numerical instability
- On prediction a categorical distribution is created to select which gaussian mixture is used per latent dimension
or overall. There is a very rare case where the pi's (mixture weights) did not sum to 1 but to 0.9999994 or sth similar
small. When then 0.9999995 or higher was drawn for the categorical distribution sample it could lead to an index
out of bounds error in the mus and sigmas. This is fixed now by making the last entry of the cumulated sum of the
mixture weights 1.0
- Additionally refactored the forward and predict functions for the GMM used in both the StandardMDN and MDNBCE as there
is no reason to have the implementation twice